BIG UPDATE
Puzzle Fun Version 1.4  Nov 9th 2013
==============================
Hi everyone!
In this release, I have simplified game play quite a bit!  Now you just
Click the start button in the game menu or click the game board to
select a puzzle.
Once  the puzzle is selected, the game will rez the tiles, and start a
count down!
If you want to End the game, click on the game menu, and select "End".
The game tiles will clean themselves up!
I also added a new "HINT" feature.  Click on the HINT button to
raise/lower a giant cube with a picture of the completed puzzle.
Also added different collision noises when the tiles fall, and cleaned
up the code quite a bit!
I hope you will have fun with this!